Rewrite CryptoExchangeFlipInput
